+++ src/components/Carousel/utils.ts
@@ -1,0 +1,12 @@
+import { useContext } from 'react';
+import { CarouselContext } from './Carousel';
+
+export function useCarousel() {
+  const context = useContext(CarouselContext);
+
+  if (!context) {
+    throw new Error('useCarousel must be used within a <Carousel />');
+  }
+
+  return context;
+}